Make sure deleteById does not contain null values
Unfortunately this library has to support old PHP versions as well, so
we are not able to use type-hints to prevent null values (for now). If
a user passes null values to certain deleteById methods this could
result in unwanted data loss, so we throw an exception to prevent any
loss.